refactor: move home components to screen folder
