+++ src/pages/index.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import * as R from 'ramda';
 import { graphql } from 'gatsby';
 
-import HomeTemplate from '../templates/home';
-import { GraphQLResponse, PostEdges } from '../types';
+import HomeTemplate from '@screens/Home/PageTemplate';
+import { GraphQLResponse, PostEdges } from 'types';
 
 const Home: React.FC<GraphQLResponse> = ({ data }) => {
   const postEdges = R.path(['allMdx', 'edges'], data) as PostEdges;
