+++ src/screens/Home/components/AuthorPresentation/index.ts
