+++ src/screens/Home/components/AuthorPresentation/styled.ts
