+++ src/screens/Home/components/Filter.tsx
@@ -3,7 +3,7 @@ import { FormattedMessage } from 'react-intl';
 import { motion } from 'framer-motion';
 
 import { css, styled } from 'styles/styled';
-import { PostFilters } from './types';
+import { PostFilters } from '@screens/Home/types';
 
 const FilterList = styled.ul`
   display: grid;
